requirements.txt
flask
tornado
requests
oslo.config
oslo.log
xmltodict
thegamesdb
pillow